requirements.txt
agno>=1.0.0
openai>=1.0.0
groq>=0.4.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0
pydantic>=2.0.0
pyjwt>=2.8.0
bcrypt>=4.1.0
requests>=2.31.0
